Bugfix in AccountSettings dialog - move <p> outside DialogContentText
Also - move AccountSettings to correct folder
